requirements.txt
playwright==1.44.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
vaderSentiment==3.3.2
keybert==0.8.4
transformers==4.41.2
torch==2.3.0
sentence-transformers==3.0.1
python-dotenv==1.0.1
httpx==0.27.0
tqdm==4.66.4
